2024 January sync

+++ shell/install_packages.sh
@@ -4,59 +4,29 @@ log_file="$(dirname "$0")/install_progress_log.txt"
 
 sudo pacman -Syuq --noconfirm
 
+download_package()  {
+    sudo pacman -S --noconfirm $1
+    if type -p $1 > /dev/null; then
+        echo "$1 installed successfully"  >> $log_file
+    else
+        echo "$1 FAILED TO INSTALL!" >> $log_file
+    fi
+}
+
+
 # -------------------
 # Productivity Tools
 # -------------------
+
 echo "Trying to install command-line utils..." >> $log_file
 
-sudo pacman -Sq --noconfirm tmux
-if type -p tmux > /dev/null; then
-    echo "tmux installed successfully"  >> $log_file
-else
-    echo "tmux FAILED TO INSTALL!" >> $log_file
-fi
-
-sudo pacman -Sq --noconfirm ranger
-if type -p ranger > /dev/null; then
-    echo "ranger intalled successfully"  >> $log_file
-else
-    echo "ranger FAILED TO INSTALL!" >> $log_file
-fi
-
-sudo pacman -Sq --noconfirm sl
-if type -p sl > /dev/null; then
-    echo "sl installed successfully"  >> $log_file
-else
-    echo "sl FAILED TO INSTALL!" >> $log_file
-fi
-
-sudo pacman -Sq --noconfirm ripgrep
-if type -p rg > /dev/null; then
-    echo "ripgrep intalled successfully"  >> $log_file
-else
-    echo "ripgrep FAILED TO INSTALL!" >> $log_file
-fi
-
-sudo pacman -Sq --noconfirm unzip
-if type -p unzip > /dev/null; then
-    echo "unzip intalled successfully"  >> $log_file
-else
-    echo "unzip FAILED TO INSTALL!" >> $log_file
-fi
-
-sudo pacman -Sq --noconfirm wget
-if type -p wget > /dev/null; then
-    echo "wget installed successfully"  >> $log_file
-else
-    echo "wget FAILED TO INSTALL!" >> $log_file
-fi
-
-sudo pacman -Sq --noconfirm curl
-if type -p curl > /dev/null; then
-    echo "curl installed successfully"  >> $log_file
-else
-    echo "curl FAILED TO INSTALL!" >> $log_file
-fi
+download_package tmux
+download_package ranger
+download_package ripgrep
+download_package unzip
+download_package wget
+download_package curl
+download_package sl
 
 echo >> $log_file
 
@@ -64,62 +34,36 @@ echo >> $log_file
 # --------------
 # Xorg and GUI
 # --------------
-echo "Trying to install GUI..." >> $log_file
 
-sudo pacman -Sq --noconfirm xorg
-if type -p Xorg > /dev/null; then
-    echo "xorg installed successfully" >> $log_file
-else
-    echo "xorg FAILED TO INSTALL!" >> $log_file
+read -n 1 -p "Install graphical environment? [y/n]" prompt
+if [[ $prompt =~ [yY] ]] then
+    echo "Trying to install GUI..." >> $log_file
+
+    download_package xorg
+    download_package xorg-xinit
+    download_package xwallpaper
+    download_package brightnessctl
+    download_package sxhkd
+
+    download_package xclip
+    download_package pamixer
+    download_package sxiv
+    download_package xournalpp
+    download_package mpv
+    download_package zathura
+    sudo pacman -Sq --noconfirm zathura-pdf-mupdf
+    sudo pacman -Sq --noconfirm imagemagick
+
+
+    echo "Trying fonts..." >> $log_file
+    sudo pacman -Sq --noconfirm ttf-jetbrains-mono
+    sudo pacman -Sq --noconfirm ttf-jetbrains-mono-nerd
+    sudo pacman -Sq --noconfirm ttf-nerd-fonts-symbols
+    sudo pacman -Sq --noconfirm ttf-nerd-fonts-symbols-mono
+    sudo pacman -Sq --noconfirm noto-fonts-cjk
+
+    echo >> $log_file
 fi
-
-sudo pacman -Sq --noconfirm xorg-xinit
-if type -p startx > /dev/null; then
-    echo "xinit installed successfully" >> $log_file
-else
-    echo "xinit FAILED TO INSTALL!" >> $log_file
-fi
-
-sudo pacman -Sq --noconfirm sxhkd
-if type -p sxhkd > /dev/null; then
-    echo "sxhkd installed successfully" >> $log_file
-else
-    echo "sxhkd FAILED TO INSTALL!" >> $log_file
-fi
-
-sudo pacman -Sq --noconfirm mpv
-if type -p mpv > /dev/null; then
-    echo "mpv installed successfully" >> $log_file
-else
-    echo "mpv FAILED TO INSTALL!" >> $log_file
-fi
-
-sudo pacman -Sq --noconfirm zathura
-if type -p zathura > /dev/null; then
-    echo "zathura installed successfully" >> $log_file
-else
-    echo "zathura FAILED TO INSTALL!" >> $log_file
-fi
-sudo pacman -Sq --noconfirm zathura-pdf-mupdf
-
-sudo pacman -Sq --noconfirm ttf-jetbrains-mono
-sudo pacman -Sq --noconfirm ttf-jetbrains-mono-nerd
-sudo pacman -Sq --noconfirm ttf-nerdfonts-symbols
-sudo pacman -Sq --noconfirm ttf-nerdfonts-symbols-mono
-
-
-if type -p yay > /dev/null; then
-    yay -Sq --noconfirm brave-bin
-    if type -p brave > /dev/null; then
-        echo "brave installed successfully" >> $log_file
-    else
-        echo "brave FAILED TO INSTALL!" >> $log_file
-    fi
-else
-    echo "yay not available, skipping brave install" >> $log_file
-fi
-
-echo >> $log_file
 
 
 # --------------
